change: Only parse for `Feature:` tags before the first Feature
Previously, we would read a Feature: tag anywhere in the file, and then
throw an exception if the file had more than one Feature.

This is different to cucumber/gherkin, which only parses a `Feature:` if
it can validly appear at that point in the document. After that:

* If free text is allowed in that part of the document (e.g. as a
  Feature or Scenario description) it will be parsed as text.
* Otherwise, it will generate the same parser error as any other
  unexpected text would.

I have therefore added multiple compatibility examples to prove this -
our existing example was valid in gherkin because the second `Feature:`
keyword just became part of the Feature description.