feat: NameFilter should match on scenario title & description for BC
When parsing in `GherkinCompatibilityMode::GHERKIN_32`, multiline text
for a Scenario (or Outline) will be split across the title and
description.

This could cause the NameFilter to unexpectedly stop matching scenarios
with multiline text if the new parsing mode has been enabled. This would
be a breaking change (and could cause false positives in CI systems if
scenarios that were assumed to be running were no longer tested).

Therefore, NameFilter is instead changed to consider both title and
description to mirror existing behaviour.

This required introducing a new interface to identify nodes that may
have a description, since `NameFilter::isScenarioMatch` accepts any
`ScenarioInterface` and that interface (as well as e.g. ExampleNode)
does not contain a `getDescription` method.

I have added that interface to all nodes - even those that are not
relevant for this feature - to avoid future confusion.